Fix compile error

+++ packages/protocol/contracts/L1/libs/LibProposing.sol
@@ -65,7 +65,7 @@ library LibProposing {
         meta.txListHash = input.txListHash;
         meta.txListByteStart = input.txListByteStart;
         meta.txListByteEnd = input.txListByteEnd;
-        meta.gasLimit = config.blockMaxGasLimit;
+        meta.gasLimit = uint32(config.blockMaxGasLimit);
         meta.beneficiary = input.beneficiary;
         meta.treasury = resolver.resolve(config.chainId, "treasury", false);
         meta.depositsProcessed =
